Assistant checkpoint: Add database message creation before websocket broadcast
Assistant generated file changes:
- server/websocket.ts: Include message ID in websocket broadcast

---

User prompt:

the issue that we have is that when we broadcast a message to a channel in the websocket.ts file, we're using the 'data' object to create the 'messageData'. But the data object (which is derived from the message that is received in the websocket's on 'message' event) only includes the following fields:

{ type: 'message', channelId: 23, content: 'f' }

this means that messageId is undefined when it received by the MessageList object, even if the message does in fact exist. therefore, because of line 50 in that file, we're simply always displaying the oldMessages which are the messages that were available when the page was originally loaded. When we need is for the msg object that gets passed to the MessageList take have its id value. How should we make this happen?